reset project

+++ src/main/java/com/sheepshop/repositorys/BillDetailRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BillDetailRepository extends JpaRepository<BillDetail, Long> {
+public interface BillDetailRepository extends JpaRepository<BillDetail, Integer> {
 
 
 }
