+++ src/main/java/com/sheepshop/repositorys/CartDetailRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CartDetailRepository extends JpaRepository<CartDetail, Long> {
+public interface CartDetailRepository extends JpaRepository<CartDetail, Integer> {
 
 
 }
