+++ src/main/java/com/sheepshop/repositorys/EmployeeRepository.java
@@ -1,5 +1,6 @@
 package com.sheepshop.repositorys;
 
+
 import com.sheepshop.entitys.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -9,7 +10,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface EmployeeRepository extends JpaRepository<Employee, Long> {
+public interface EmployeeRepository extends JpaRepository<Employee, Integer> {
 
     @Query(value = "Select e from Employee e where e.status = 0")
     public List<Employee> getAll();
@@ -18,13 +19,13 @@ public interface EmployeeRepository extends JpaRepository<Employee, Long> {
     public List<Employee> searchByName(@Param("fullname") String fullname);
 
     @Query(value = "select e from Employee e where e.id = :id")
-    public Employee getById(@Param("id") Long Id);
+    public Employee getById(@Param("id") Integer Id);
 
     @Query(value = "select e from Employee e where e.username = :username")
     public Employee getByUsername(@Param("username") String username);
 
     @Query(value = "select e from Employee e where e.status = 0 and (e.role.id = :idRole or :idRole is null)")
-    public List<Employee> getEmployeeByRole(@Param("idRole") Long id);
+    public List<Employee> getEmployeeByRole(@Param("idRole") Integer id);
 
     @Query(value = "Select e from Employee e where e.status = :status")
     public List<Employee> getByStatus(@Param("status") Integer status);
